pyproject.toml
[project]
name = "tsmc-2026-hackathon"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "mypy>=1.19.1",
    "pathspec>=0.12.1",
    "pydantic>=2.11.7",
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
    "google-generativeai>=0.8.6",
    "google-cloud-aiplatform>=1.135.0",
]

# -------------------------
# Ruff (lint + format)
# -------------------------
[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# -------------------------
# Mypy
# -------------------------
[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
